Refactor to Abstract
