--- cmd/assistant/pkg/assistant/task-based-constants.go
@@ -1,11 +1,0 @@
-// Copyright 2022. All Rights Reserved.
-// SPDX-License-Identifier: MIT
-
-package assistantimpl
-
-import "errors"
-
-var (
-	// ErrNoActiveTask no active task
-	ErrNoActiveTask = errors.New("no active task")
-)
