+++ pkg/microphone/microphone.go
@@ -1,3 +1,7 @@
+// Copyright 2023 The dvonthenen Open-Virtual-Assistant Authors. All Rights Reserved.
+// Use of this source code is governed by an Apache-2.0 license that can be found in the LICENSE file.
+// SPDX-License-Identifier: Apache-2.0
+
 package microphone
 
 import (
