+++ pkg/speech/speech.go
@@ -1,5 +1,6 @@
-// Copyright 2022. All Rights Reserved.
-// SPDX-License-Identifier: MIT
+// Copyright 2023 The dvonthenen Open-Virtual-Assistant Authors. All Rights Reserved.
+// Use of this source code is governed by an Apache-2.0 license that can be found in the LICENSE file.
+// SPDX-License-Identifier: Apache-2.0
 
 package speech
 
@@ -16,6 +17,8 @@ import (
 	"github.com/faiface/beep/speaker"
 	texttospeechpb "google.golang.org/genproto/googleapis/cloud/texttospeech/v1"
 	klog "k8s.io/klog/v2"
+
+	interfaces "github.com/dvonthenen/open-virtual-assistant/pkg/speech/interfaces"
 )
 
 type SpeechOptions struct {
@@ -33,10 +36,10 @@ func New(ctx context.Context, opts *SpeechOptions) (*Client, error) {
 	klog.V(6).Infof("speech.New ENTER\n")
 
 	if opts.LanguageCode == "" {
-		opts.LanguageCode = DefaultLanguageCode
+		opts.LanguageCode = interfaces.DefaultLanguageCode
 	}
 	if opts.VoiceType == 0 {
-		opts.VoiceType = SpeechVoiceNeutral
+		opts.VoiceType = interfaces.SpeechVoiceNeutral
 	}
 
 	var googleCredentials string
