+++ pkg/transcriber/deepgram/deepgram.go
@@ -1,5 +1,6 @@
-// Copyright 2022. All Rights Reserved.
-// SPDX-License-Identifier: MIT
+// Copyright 2023 The dvonthenen Open-Virtual-Assistant Authors. All Rights Reserved.
+// Use of this source code is governed by an Apache-2.0 license that can be found in the LICENSE file.
+// SPDX-License-Identifier: Apache-2.0
 
 package deepgram
